docs: k6에 대한 내용 업데이트
- Scenarios 에 대한 내용 및 그림 추가
