+++ playground/devfancy/spring-performance/k6-scripts/first_scripts.js
@@ -1,20 +1,7 @@
 import http from 'k6/http';
 import { sleep } from 'k6';
 
-export const options = {
-    discardResponseBodies: true,
-    scenarios: {
-        oom: {
-            executor: 'constant-vus',
-            vus: 50,
-            duration: '60s',
-        }
-    },
-    noConnectionReuse: true,
-};
-
 export default function () {
-    const testUrl = 'http://localhost:8080/api/external/countries';
-    http.get(testUrl);
+    http.get('https://test.k6.io');
     sleep(1);
 }
